// src/generated/models/Cgmp_bridgesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cgmp_bridgescgmp_status = {
  100000000: 'Active',
  100000001: 'Completed',
  100000002: 'Failed',
  100000003: 'Cancelled',
  100000004: 'Scheduled'
} as const;
export type Cgmp_bridgescgmp_status = keyof typeof Cgmp_bridgescgmp_status;
export const Cgmp_bridgescgmp_template = {
  100000000: 'Network',
  100000001: 'Citrix',
  100000002: 'VMware',
  100000003: 'Firewall',
  100000004: 'PatchActivities',
  100000005: 'Custom'
} as const;
export type Cgmp_bridgescgmp_template = keyof typeof Cgmp_bridgescgmp_template;
export const Cgmp_bridgesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cgmp_bridgesstatecode = keyof typeof Cgmp_bridgesstatecode;
export const Cgmp_bridgesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cgmp_bridgesstatuscode = keyof typeof Cgmp_bridgesstatuscode;

export interface Cgmp_bridgesBase {
  cgmp_actualend?: string;
  cgmp_actualstart?: string;
  cgmp_bridgeid: string;
  cgmp_changeid?: string;
  cgmp_changenumber?: string;
  cgmp_closureremarks?: string;
  cgmp_downtimeminutes?: number;
  cgmp_duration?: number;
  cgmp_lessonslearned?: string;
  cgmp_participants?: string;
  cgmp_pirnotes?: string;
  cgmp_projectstatuses?: string;
  cgmp_rollbackinitiated?: boolean;
  cgmp_rollbackreason?: string;
  cgmp_schedend: string;
  cgmp_schedstart: string;
  cgmp_status: Cgmp_bridgescgmp_status;
  cgmp_teamsurl?: string;
  cgmp_template: Cgmp_bridgescgmp_template;
  cgmp_title: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cgmp_bridgesstatecode;
  statuscode?: Cgmp_bridgesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cgmp_bridges extends Cgmp_bridgesBase {
  cgmp_rollbackinitiatedname?: string;
  cgmp_statusname?: string;
  cgmp_templatename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}